feat: Add solution to the problem Three Activities
This commit adds the solution to the problem "Three Activities" on Codeforces. The solution is implemented in the file "1914d_ta.cpp". It uses a set and vectors to store and sort data, and calculates the maximum sum of three elements from three different vectors. The solution passes all test cases.