feat: Add solution to the problem Romantic Glasses
This commit adds the solution to the problem Romantic Glasses. The code implements a solution that checks if there are two subarrays with equal sums in a given array. The solution uses prefix sums and sorting to efficiently find the answer. The code has been tested and is ready for use.