feat: Add solution to the problem Monsters
This commit adds the solution to the problem Monsters. The code implements a sorting algorithm to sort a list of numbers based on their remainders when divided by a given number. The sorted list is then printed. The solution is based on the problem statement provided in the Codeforces platform.